renamed config folder
